+++ src/core/mod.rs
@@ -1,3 +1,5 @@
 mod event_target_maybe_signal;
+mod position;
 
 pub use event_target_maybe_signal::*;
+pub use position::*;