+++ src/core/position.rs
@@ -1,0 +1,5 @@
+#[derive(Copy, Clone)]
+pub struct Position {
+    pub x: f64,
+    pub y: f64,
+}