+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - `use_supported`
 - `use_resize_observer`
 - `watch`
+- `use_mouse`
+
+#### Other Changes
+- Use the crate `default-struct-builder` to provide ergonimic function options.
 
 ## 0.1.3